Add active class to the link element for the blog when the current page is blog-index.html.
In the navbar.html partial, the link is added to the appropriate li
element if the current page is equal to "/blog-index.html".

Because the vitality.js script invokes Bootstrap's ScrollSpy plugin, and
the plugin would remove the active class from the li element, it was
necessary to modify vitality.js to only invoke the ScrollSpy plugin if
the current page is "/index.html".